feat(dre): Improve the summary of the subnet resize NNS proposals
In addition to the improved summary:

Changes:
- arguments --add and --remove now include default values, allowing users to omit these values if not needed.
-- the `id` field has an alias of `subnet-id` for improved readability.

Refactor: the `execute` method now uses the `subnet_manager` to resize the subnet and proposes the subnet change using the `runner`. This makes the `resize` functionality very similar to the `replace` functionality, so more code can be shared between the two.
